Se agregan validaciones para campos vacios y cuando la tarjeta no esta colocada para escribir

+++ app/src/main/java/com/example/vistaspruebas/TarjetaResponse.kt
@@ -87,7 +87,7 @@ data class Subsidio(
     val id: String,
     val clave: String,
     val nombre: String,
-    val diasUtiles: Long,
+    val diasUtiles: Int,
 )
